refactor: move conversion to hashed Merkle proof into tree crate
