+++ crates/tree/Cargo.toml
@@ -11,15 +11,25 @@ keywords.workspace = true
 readme.workspace = true
 
 [dependencies]
-jmt.workspace = true
-sha2.workspace = true
-serde.workspace = true
-log.workspace = true
-anyhow.workspace = true
+# prism
 prism-common = { workspace = true, features = ["test_utils"] }
 prism-errors.workspace = true
 prism-keys.workspace = true
+
+# tree-impl
+jmt.workspace = true
+sha2.workspace = true
+
+# serde
 prism-serde.workspace = true
+serde.workspace = true
+
+# OAS spec
+utoipa.workspace = true
+
+# misc
+anyhow.workspace = true
+log.workspace = true
 
 [dev-dependencies]
 paste = { workspace = true }
